Format and sort availability display
Previously the ui was not sorting the users available days for a topic on their Account page and was displaying the days in the order they were clicked on the application page.
The ui will now take the availability from the db, match those entries to a 'enum' like structure, and sort them before displaying.